pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hmsim"
version = "0.1.0"
description = "Multi-version simulator for HM 16-bit processor family"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Khairulmizam Samsudin", email = "xource@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

gui = [
    "PyGObject>=3.44",
    "markdown-it-py>=3.0.0",
]

all = [
    "PyGObject>=3.44",
    "markdown-it-py>=3.0.0",
    "pytest>=7.0",
]

[project.scripts]
hmsim_cli = "hmsim.tools.hmsim_cli:main"
hmsim = "hmsim.gui.hm_gui:main"
hmasm = "hmsim.tools.hmasm:main"
hmdas = "hmsim.tools.hmdas:main"

[tool.setuptools.packages.find]
where = ["src"]
